supplier_research_app/requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.3
bcrypt>=4.0.0
anthropic>=0.18.0
httpx>=0.27.0
python-dotenv>=1.0.0
